# server/api/openapi-with-docs.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Dead Pigeons API",
    "description": "API for the Dead Pigeons bingo project.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Board/{id}": {
      "get": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updatedBoard",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Board"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Transaction": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "Transaction_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Transaction"
        ],
        "operationId": "Transaction_Create",
        "requestBody": {
          "x-name": "transaction",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Transaction/{id}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "Transaction_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Transaction"
        ],
        "operationId": "Transaction_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updatedTransaction",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transaction"
        ],
        "operationId": "Transaction_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/WinningBoard": {
      "get": {
        "tags": [
          "WinningBoard"
        ],
        "operationId": "WinningBoard_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "WinningBoard"
        ],
        "operationId": "WinningBoard_Create",
        "requestBody": {
          "x-name": "board",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Winningboard"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/WinningBoard/{id}": {
      "get": {
        "tags": [
          "WinningBoard"
        ],
        "operationId": "WinningBoard_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "WinningBoard"
        ],
        "operationId": "WinningBoard_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updatedBoard",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Winningboard"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "WinningBoard"
        ],
        "operationId": "WinningBoard_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Board": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "boardId": {
            "type": "string",
            "format": "guid"
          },
          "playerId": {
            "type": "string",
            "format": "guid"
          },
          "gameId": {
            "type": "string",
            "format": "guid"
          },
          "chosenNumbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "price": {
            "type": "number",
            "format": "decimal"
          },
          "isRepeating": {
            "type": "boolean"
          },
          "repeatUntilGameId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "player": {
            "$ref": "#/components/schemas/Player"
          },
          "repeatUntilGame": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Game"
              }
            ]
          },
          "winningboards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Winningboard"
            }
          }
        }
      },
      "Game": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "gameId": {
            "type": "string",
            "format": "guid"
          },
          "winningNumbers": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "drawDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time"
          },
          "boardGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          },
          "boardRepeatUntilGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          },
          "winningboards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Winningboard"
            }
          }
        }
      },
      "Winningboard": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "winningboardId": {
            "type": "string",
            "format": "guid"
          },
          "gameId": {
            "type": "string",
            "format": "guid"
          },
          "boardId": {
            "type": "string",
            "format": "guid"
          },
          "winningNumbersMatched": {
            "type": "integer",
            "format": "int32"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "board": {
            "$ref": "#/components/schemas/Board"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          }
        }
      },
      "Player": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "playerId": {
            "type": "string",
            "format": "guid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "passwordHash": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "boards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        }
      },
      "Transaction": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "transactionId": {
            "type": "string",
            "format": "guid"
          },
          "playerId": {
            "type": "string",
            "format": "guid"
          },
          "mobilepayReqId": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "status": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "player": {
            "$ref": "#/components/schemas/Player"
          }
        }
      }
    }
  }
}